layout and style of navigation
