+++ src/components/Nav/SubMenu.jsx
@@ -3,7 +3,7 @@ import LinkButton from "./LinkButton";
 import SubMenuControl from "./SubMenuControl";
 
 const SubLink = ({sublink, menuDispatch}) => (
-  <LinkButton menuDispatch={menuDispatch} styleAttribute="bg-red-100">
+  <LinkButton menuDispatch={menuDispatch} styleAttribute="bg-red-100 h-16 w-40">
     <LinkRoute
       linkTitle={sublink.title}
       path={sublink.path}
@@ -21,7 +21,7 @@ export default function SubMenu({menuDispatch, subroutes, subRouteView}) {
       routeTitle={routeTitle}
       subRouteView={subRouteView}
     >
-      <div className="flex flex-col absolute top-10 right-0">
+      <div className="flex flex-col absolute -right-5 top-10">
         {routes.map((sublink) => (
           <SubLink
             key={sublink.id}
